Authorization Prototype
The first prototype for using an authorization helper daemon. The code
heavily borrows from Apple's EvenBetterAuthorizationSample[1]. It installs
the helper tool via SMJobBless. This requires Developer ID code signing.
Communication with the helper tool is established via NSXPCConnection.
Installation and communication works, but the logic for when to establish
a connection and when to install the helper is not finalized, and the helper
code is not implemented to actually perform any meaningful action.
[1] https://developer.apple.com/library/content/samplecode/EvenBetterAuthorizationSample/Introduction/Intro.html