ZFS Error Reporting from libzfs
Expose the libzfs error strings via the API and use it to signal errors
from the helper. It is unclear how helpful this is in general, but it is
likely that those errors contain more detail than the error codes in most
situations.
This error string is now used to signal more detailed mount / unmount /
export / import errors.