Helper Proxy communication threading bug fixes
Attempt to solve the proxy threading issues at the root, by dispatching
all replies to the main thread instead of leaving this to the individual
consumers. This should be less error prone and make it harder to forget
the main thread dispatching. In this program, there's little to be gained
by not dispatching to the main thread, since almost every reply needs to
interact with the UI anyway.