Snapshot Creation / Cloning / Destruction
Snapshots can now be created from ZetaWatch, for now with only a fixed
name.Snapshots can now be cloned, the name of the clone is automatically
determined based on the snapshot name. Snapshots and file systems can now
be destroyed, but no recursive destruction is available, and error
checking is minimal.