Mount / Destroy Recursive Cleanup
Mount and Automount now use public check functions and succeed if
there is nothing to do (if canmount=off or not =on). This makes the
failure reason the same as with the zfs command line tool. Destroy
on snapshots is now only shown if it is permissible (=no clones),
otherwise, only Destroy Recursive is shown.