ZPool Import Altroot Support
Setting an altroot is now supported in the ZFSWrapper, the Authorization
Helper and the GUI. A new settings dialog contains the setting of altroot,
since this was troublesome to implement inside the menu.
Altroot, if it is enabled, is passed to all imports, including auto imports.
There is currently no way to selectively enable altroots only for certain
pools.