Filesystem / Volume create split
Split creation of filesystems and volumes in the helper protocol. The two
functionalities are separate in the frontend and backend due to different
requirements, so feeding them through the same channel makes little sense.
It is a single function (zfs_create) in the zfs library, but this is not
exposed at all.
The destroyFilesystem method is now called destroy, since it destroys not
just filesystems but also volumes, snapshots and bookmarks.