Search Path Override
Allow overriding of the search path when importing pools, or searching for
importable pools. This is controlled by a single setting, "searchPathOverride".
Currently, the value of this setting defaults to by-serial and by-id, and can
be changed by editing the settings file. The GUI only shows the current value,
modification is not implemented.